+++ CHANGELOG.md
@@ -39,3 +39,4 @@ your changes, such as:
 - [public] [both] [doc] add ilogtail local deloyment doc for sls flusher
 - [public] [both] [doc] add ilogtail k8s deloyment doc for sls flusher 
 - [public] [both] [bug] fix wrong matching lables for stdout input
+- [public] [both] [doc] add input stdout plugin doc